Move benchmarks to benches module
This ensures that the benchmarks are in a separate crate and linked
against the smallvec dynamic library rather than being compiled
together.
